# Fix performance of group#show 