--- a/src/api/app/models/bs_request/find_for/group.rb
+++ b/src/api/app/models/bs_request/find_for/group.rb
@@ -8,14 +8,14 @@ class BsRequest
         group = ::Group.find_by_title!(group_title)
 
         # find requests where group is maintainer in target project
-        @relation, inner_or = extend_query_for_maintainer(group, @relation, roles, inner_or)
+        inner_or = extend_query_for_maintainer(group, roles, inner_or)
 
         if roles.empty? || roles.include?('reviewer')
           @relation = @relation.includes(:reviews).references(:reviews)
           # requests where the user is reviewer or own requests that are in review by someone else
           or_in_and = ["reviews.by_group=#{quote(group.title)}"]
 
-          @relation, inner_or = extend_query_for_involved_reviews(group, or_in_and, @relation, review_states, inner_or)
+          inner_or = extend_query_for_involved_reviews(group, or_in_and, review_states, inner_or)
         end
         if inner_or.empty?
           @relation.none