--- a/src/api/app/models/bs_request/find_for/group.rb
+++ b/src/api/app/models/bs_request/find_for/group.rb
@@ -4,24 +4,39 @@ class BsRequest
       include UserGroupMixin
 
       def all
-        inner_or = []
-        group = ::Group.find_by_title!(group_title)
+        query = union_query
+        @relation = if query.present?
+                      @relation.where("bs_requests.id IN (#{query})")
+                    else
+                      @relation.none
+                    end
+      end
 
-        # find requests where group is maintainer in target project
-        inner_or = extend_query_for_maintainer(group, roles, inner_or)
+      def all_count
+        ActiveRecord::Base.connection.execute("SELECT COUNT(bs_request_id) FROM (#{union_query}) x").first.first
+      end
 
-        if roles.empty? || roles.include?('reviewer')
-          @relation = @relation.includes(:reviews).references(:reviews)
-          # requests where the user is reviewer or own requests that are in review by someone else
-          or_in_and = ["reviews.by_group=#{quote(group.title)}"]
+      private
 
-          inner_or = extend_query_for_involved_reviews(group, or_in_and, review_states, inner_or)
-        end
-        if inner_or.empty?
-          @relation.none
-        else
-          @relation.where(inner_or.join(' or '))
-        end
+      def group
+        @group ||= ::Group.find_by_title!(group_title)
+      end
+
+      def union_query
+        query_parts = []
+        query_parts << bs_request_actions_query.to_sql if bs_request_actions_query
+        query_parts << reviews_query.to_sql if reviews_query
+        query_parts.join(' UNION ')
+      end
+
+      def bs_request_actions_query
+        return unless roles.empty? || roles.include?('maintainer')
+        @bs_request_actions_query ||= bs_request_actions(group).select(:bs_request_id)
+      end
+
+      def reviews_query
+        return unless roles.empty? || roles.include?('reviewer')
+        @reviews_query ||= reviews(group, review_states).select(:bs_request_id)
       end
     end
   end