--- a/src/api/app/views/webui2/webui/groups/show.html.haml
+++ b/src/api/app/views/webui2/webui/groups/show.html.haml
@@ -24,7 +24,7 @@
         %a.nav-link.text-nowrap#all-requests-tab{ aria: { controls: 'all-requests' }, data: { toggle: 'tab' }, href: '#all-requests', role: 'tab' }
           All Requests
           %span.badge.badge-primary
-            = @group.requests.size
+            = @group.all_requests_count
       %li.nav-item.dropdown
         %a.nav-link.dropdown-toggle{ href: '#', 'data-toggle': 'dropdown', 'role': 'button', 'aria-expanded': 'false', 'aria-haspopup': 'true' }
         .dropdown-menu.dropdown-menu-right