--- a/src/api/db/migrate/20190412130831_add_state_group_index_to_reviews.rb
+++ b/src/api/db/migrate/20190412130831_add_state_group_index_to_reviews.rb
@@ -1,0 +1,5 @@
+class AddStateGroupIndexToReviews < ActiveRecord::Migration[5.2]
+  def change
+    add_index :reviews, [:state, :by_group]
+  end
+end