The following call done in the group#show page is extremely slow (> 20 seconds):

```
BsRequest::FindFor::Group.new(params).all
```

because of the complex query that it generates. Refactor how the query is generated, writing in the Rail way improving the performance.

We now use the UNION feature of mysql, which allows to avoid the bottleneck of OR avoiding indices. The UNION joins 2 SELECT calls that are technically a OR of a join, but simplifies the way the query optimizer can use indices.

The same applies to FindFor::User, but this code is only used when calling the API with /request?view=collection&user=X.

benchmarks (on production clone of January, just counting the rails time without javascript):

wicked-maintainers:  7077ms -> 312ms
opensuse-review-team: 834ms -> 552ms
gnome-maintainers: 13885ms -> 1889ms

(gnome-maintainers is still a bit slow as it has a lot of reviews and the reviews trouble mariadb indexes, but we decided to stop here).

Fixes #7139

Co-authored-by: @Ana06
Co-authored-by: @andrii-suse