@Ana06 2019-04-16
![](https://github.trello.services/images/mini-trello-icon.png) [[Last card] Address feedback/issues](https://trello.com/c/9QtsFies/1097-last-card-address-feedback-issues)