@Ana06 2019-04-16
I think that Hakiri warnings are all false positives. @openSUSE/open-build-service does anybody think differently? Please review :wink: