@andrii-suse 2019-04-16
> I think that Hakiri warnings are all false positives.
They would be true positives if user input would go into them directly. As long as those subqueries (union_query, users_query and packages_query) are safe - result query will be safe as well, so it should be safe to ignore the warnings.